Выносит в контроллер отрисовку формы
